4-scan_ai-Manus/backend/tests/unit/test_auth_api.py
"""
اختبارات وحدة API المصادقة
Authentication API Unit Tests
"""

import pytest
from fastapi.testclient import TestClient
from unittest.mock import patch, MagicMock
import jwt
from datetime import datetime, timedelta

from src.main import app
from src.core.config import get_settings

client = TestClient(app)
settings = get_settings()

class TestAuthAPI:
    """اختبارات API المصادقة"""

    def test_successful_login(self):
        """اختبار تسجيل دخول ناجح"""
        login_data = {
            "username": "test@gaara.ai",
            "password": "TestPassword123!"
        }

        response = client.post("/api/v1/auth/login", json=login_data)

        assert response.status_code == 200
        data = response.json()

        assert "access_token" in data
        assert data["token_type"] == "bearer"
        assert data["expires_in"] == settings.JWT_EXPIRATION_HOURS * 3600

        # التحقق من صحة التوكن
        token = data["access_token"]
        payload = jwt.decode(token, settings.JWT_SECRET, algorithms=[settings.JWT_ALGORITHM])
        assert "sub" in payload  # sub contains user ID
        assert "email" in payload

    def test_failed_login_wrong_password(self):
        """اختبار تسجيل دخول فاشل - كلمة مرور خاطئة"""
        login_data = {
            "username": "test@gaara.ai",
            "password": "wrong_password"
        }

        response = client.post("/api/v1/auth/login", json=login_data)

        assert response.status_code == 401
        data = response.json()

        # Response can have detail or message field
        error_msg = data.get("detail", data.get("message", ""))
        assert "invalid" in error_msg.lower() or "credentials" in error_msg.lower()

    def test_failed_login_wrong_username(self):
        """اختبار تسجيل دخول فاشل - اسم مستخدم خاطئ"""
        login_data = {
            "username": "wrong_user@gaara.ai",
            "password": "TestPassword123!"
        }

        response = client.post("/api/v1/auth/login", json=login_data)

        assert response.status_code == 401
    
    def test_login_missing_fields(self):
        """اختبار تسجيل دخول مع حقول ناقصة"""
        # بدون كلمة مرور
        response = client.post("/api/v1/auth/login", json={"username": "admin"})
        assert response.status_code == 422
        
        # بدون اسم مستخدم
        response = client.post("/api/v1/auth/login", json={"password": "admin123"})
        assert response.status_code == 422
        
        # بدون أي حقول
        response = client.post("/api/v1/auth/login", json={})
        assert response.status_code == 422
    
    def test_get_current_user_valid_token(self):
        """اختبار الحصول على المستخدم الحالي - توكن صالح"""
        # تسجيل دخول أولاً
        login_data = {"username": "test@gaara.ai", "password": "TestPassword123!"}
        login_response = client.post("/api/v1/auth/login", json=login_data)
        if login_response.status_code != 200:
            pytest.skip("Test user not found in database")
        token = login_response.json()["access_token"]

        # الحصول على معلومات المستخدم
        headers = {"Authorization": f"Bearer {token}"}
        response = client.get("/api/v1/auth/me", headers=headers)

        assert response.status_code == 200
        data = response.json()

        # Just verify it has email field, not specific value
        assert "email" in data

    def test_get_current_user_invalid_token(self):
        """اختبار الحصول على المستخدم الحالي - توكن غير صالح"""
        headers = {"Authorization": "Bearer invalid_token"}
        response = client.get("/api/v1/auth/me", headers=headers)

        assert response.status_code == 401

    def test_get_current_user_expired_token(self):
        """اختبار الحصول على المستخدم الحالي - توكن منتهي الصلاحية"""
        # إنشاء توكن منتهي الصلاحية
        payload = {
            "sub": 1,  # user ID
            "email": "test@gaara.ai",
            "exp": datetime.utcnow() - timedelta(hours=1),  # منتهي منذ ساعة
            "iat": datetime.utcnow() - timedelta(hours=2),
        }
        expired_token = jwt.encode(payload, settings.JWT_SECRET, algorithm=settings.JWT_ALGORITHM)

        headers = {"Authorization": f"Bearer {expired_token}"}
        response = client.get("/api/v1/auth/me", headers=headers)

        assert response.status_code == 401

    def test_get_current_user_no_token(self):
        """اختبار الحصول على المستخدم الحالي - بدون توكن"""
        response = client.get("/api/v1/auth/me")

        assert response.status_code == 403  # Forbidden

    def test_logout_success(self):
        """اختبار تسجيل خروج ناجح"""
        # تسجيل دخول أولاً
        login_data = {"username": "test@gaara.ai", "password": "TestPassword123!"}
        login_response = client.post("/api/v1/auth/login", json=login_data)
        if login_response.status_code != 200:
            pytest.skip("Test user not found in database")
        token = login_response.json()["access_token"]

        # تسجيل خروج
        headers = {"Authorization": f"Bearer {token}"}
        response = client.post("/api/v1/auth/logout", headers=headers)

        assert response.status_code == 200

    def test_logout_no_token(self):
        """اختبار تسجيل خروج بدون توكن - يرجع 200 أو 403"""
        response = client.post("/api/v1/auth/logout")

        # Logout without token can succeed (clears cookies) or fail (requires auth)
        assert response.status_code in [200, 403]

class TestAuthAPIValidation:
    """اختبارات التحقق من صحة البيانات"""
    
    def test_login_empty_username(self):
        """اختبار تسجيل دخول مع اسم مستخدم فارغ"""
        login_data = {"username": "", "password": "admin123"}
        response = client.post("/api/v1/auth/login", json=login_data)
        
        assert response.status_code == 401  # سيفشل في التحقق
    
    def test_login_empty_password(self):
        """اختبار تسجيل دخول مع كلمة مرور فارغة"""
        login_data = {"username": "admin", "password": ""}
        response = client.post("/api/v1/auth/login", json=login_data)
        
        assert response.status_code == 401  # سيفشل في التحقق
    
    def test_login_whitespace_credentials(self):
        """اختبار تسجيل دخول مع مسافات فقط"""
        login_data = {"username": "   ", "password": "   "}
        response = client.post("/api/v1/auth/login", json=login_data)
        
        assert response.status_code == 401

class TestAuthAPISecurity:
    """اختبارات الأمان"""

    def test_token_contains_no_sensitive_data(self):
        """اختبار أن التوكن لا يحتوي على بيانات حساسة"""
        login_data = {"username": "test@gaara.ai", "password": "TestPassword123!"}
        response = client.post("/api/v1/auth/login", json=login_data)
        if response.status_code != 200:
            pytest.skip("Test user not found in database")
        token = response.json()["access_token"]

        # فك تشفير التوكن
        payload = jwt.decode(token, settings.JWT_SECRET, algorithms=[settings.JWT_ALGORITHM])

        # التحقق من عدم وجود كلمة المرور
        assert "password" not in payload
        assert "TestPassword123!" not in str(payload)

        # التحقق من وجود البيانات المطلوبة فقط (sub, exp, iat)
        required_fields = {"sub", "exp", "iat"}
        assert required_fields.issubset(set(payload.keys()))

    def test_multiple_login_attempts(self):
        """اختبار محاولات تسجيل دخول متعددة"""
        login_data = {"username": "nonexistent@gaara.ai", "password": "wrong_password"}

        # محاولة 5 مرات
        for _ in range(5):
            response = client.post("/api/v1/auth/login", json=login_data)
            assert response.status_code == 401

        # التحقق من أن النظام لا يزال يستجيب
        correct_login = {"username": "test@gaara.ai", "password": "TestPassword123!"}
        response = client.post("/api/v1/auth/login", json=correct_login)
        # Test user may not exist, so check for success or skip
        assert response.status_code in [200, 401]

