4-scan_ai-Manus/backend/requirements.txt
# FILE: backend/requirements.txt | PURPOSE: Backend dependencies | OWNER: Backend Team | LAST-AUDITED: 2025-12-13
# Gaara AI Backend Dependencies - Canonical Version 3.0
# Merged from: src/requirements.txt, clean_project/requirements.txt, gaara_ai_integrated/backend/requirements.txt

# ==========================================
# Core Framework
# ==========================================
fastapi==0.115.12
uvicorn[standard]==0.34.3
pydantic==2.11.7
pydantic-settings==2.1.0
python-dotenv==1.1.0

# ==========================================
# Database
# ==========================================
sqlalchemy==2.0.23
psycopg2-binary==2.9.10
alembic==1.13.1

# ==========================================
# Redis and Caching
# ==========================================
redis==5.0.1
aioredis==2.0.1

# ==========================================
# Security and Authentication
# ==========================================
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pyotp>=2.9.0
qrcode[pil]>=7.4.2
bcrypt==4.1.2
cryptography==45.0.4
PyJWT==2.10.1
pyotp==2.9.0
qrcode==7.4.2
pyOpenSSL==25.1.0
bleach==6.1.0
slowapi==0.1.9

# ==========================================
# Image Processing
# ==========================================
Pillow==11.2.1
opencv-python==4.8.1.78
numpy==1.26.4
scikit-image==0.22.0

# ==========================================
# AI and Machine Learning
# ==========================================
torch==2.1.1
tensorflow==2.15.0
scikit-learn==1.3.2
onnx==1.18.0
onnx-tf==1.10.0
coremltools==8.3.0; platform_system == "Darwin"

# ==========================================
# Data Processing
# ==========================================
pandas==2.3.0
matplotlib==3.10.3
openpyxl==3.1.5

# ==========================================
# Web Framework (Flask - for compatibility)
# ==========================================
Flask==3.1.1
flask-cors==6.0.0
flask-jwt-extended==4.7.1
flask-login==0.6.3
flask-sqlalchemy==3.1.1

# ==========================================
# PDF Generation
# ==========================================
fpdf2==2.8.3
pdfkit==1.0.0

# ==========================================
# Monitoring and Logging
# ==========================================
prometheus-client==0.22.1
psutil==7.0.0
evidently==0.7.7
langfuse==2.60.10

# ==========================================
# Testing
# ==========================================
pytest==8.3.5
pytest-asyncio==0.21.1
pytest-cov==4.1.0
playwright==1.52.0
httpx==0.25.2

# ==========================================
# Utilities
# ==========================================
Jinja2==3.1.6
filelock==3.18.0
protobuf==4.25.3
python-magic==0.4.27
Cython==3.1.2
docutils==0.21.2
keyring==25.6.0
ipython==8.12.3
ipywidgets==8.1.7

# Tensorflow requires wrapt<1.15; langfuse allows >=1.14
wrapt==1.14.2

# ==========================================
# Async and Task Queue
# ==========================================
celery==5.3.4
kombu==5.3.4
amqp==5.2.0

# ==========================================
# Email
# ==========================================
email-validator==2.1.0
aiosmtplib==3.0.1

# ==========================================
# HTTP Clients
# ==========================================
requests==2.32.3
aiohttp==3.9.1

# aiohttp dependency (pin to reduce pip resolver backtracking)
multidict==6.7.0

# ==========================================
# Development Tools
# ==========================================
black==23.12.1
flake8==6.1.0
isort==5.13.2
mypy==1.7.1
pre-commit==3.6.0

# ==========================================
# Documentation
# ==========================================
mkdocs==1.5.3
mkdocs-material==9.5.3

# ==========================================
# Environment Management
# ==========================================
python-decouple==3.8
environs==10.3.0

# ==========================================
# Date and Time
# ==========================================
python-dateutil==2.8.2
pytz==2023.3

# ==========================================
# Validation
# ==========================================
phonenumbers==8.13.27

# ==========================================
# File Handling
# ==========================================
filetype==1.2.0

# ==========================================
# Websockets
# ==========================================
websockets==12.0
python-socketio==5.10.0

# ==========================================
# Message Queue
# ==========================================
pika==1.3.2


# Data Management & Scheduling
apscheduler==3.10.4
beautifulsoup4==4.12.2
